refactor: lift the pure playlist request helpers out of the hook
